requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
Brotli==1.1.0
celery==5.4.0
cffi==1.17.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cssselect2==0.7.0
Django==5.1
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
fonttools==4.53.1
html5lib==1.1
inflection==0.5.1
kombu==5.4.0
mysqlclient==2.2.4
packaging==24.1
pillow==10.4.0
prompt_toolkit==3.0.47
pycparser==2.22
pydyf==0.11.0
PyJWT==2.9.0
pyphen==0.16.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
redis==5.0.8
six==1.16.0
sqlparse==0.5.1
tinycss2==1.3.0
tzdata==2024.1
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
weasyprint==62.3
webencodings==0.5.1
zopfli==0.2.3
